refactor: id, createdAt, updatedAt 필드 분리
